// src/schema/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * A date string, such as 2007-12-03, compliant with the `full-date` format
   * outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: { input: any; output: any; }
  /**
   * A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the
   * `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO
   * 8601 standard for representation of dates and times using the Gregorian calendar.
   */
  DateTime: { input: any; output: any; }
  /** A string used to identify an i18n locale */
  I18NLocaleCode: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
  /** The `Upload` scalar type represents a file upload. */
  Upload: { input: any; output: any; }
};
// ...
export type I18NLocaleEntityResponse = {
  __typename?: 'I18NLocaleEntityResponse';
  data?: Maybe<I18NLocaleEntity>;
};

export type I18NLocaleEntityResponseCollection = {
  __typename?: 'I18NLocaleEntityResponseCollection';
  data: Array<I18NLocaleEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type IgloohomeChangelog = {
  __typename?: 'IgloohomeChangelog';
  AppVersion?: Maybe<Scalars['String']['output']>;
  Changes?: Maybe<Scalars['String']['output']>;
  ReleaseDate?: Maybe<Scalars['Date']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  locale?: Maybe<Scalars['String']['output']>;
  localizations?: Maybe<IgloohomeChangelogRelationResponseCollection>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};


export type IgloohomeChangelogLocalizationsArgs = {
  filters?: InputMaybe<IgloohomeChangelogFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type IgloohomeChangelogEntity = {
  __typename?: 'IgloohomeChangelogEntity';
  attributes?: Maybe<IgloohomeChangelog>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type IgloohomeChangelogEntityResponse = {
  __typename?: 'IgloohomeChangelogEntityResponse';
  data?: Maybe<IgloohomeChangelogEntity>;
};

export type IgloohomeChangelogEntityResponseCollection = {
  __typename?: 'IgloohomeChangelogEntityResponseCollection';
  data: Array<IgloohomeChangelogEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type IgloohomeChangelogRelationResponseCollection = {
  __typename?: 'IgloohomeChangelogRelationResponseCollection';
  data: Array<IgloohomeChangelogEntity>;
};

export type IgloohomeFaq = {
  __typename?: 'IgloohomeFaq';
  Category?: Maybe<IgloohomeFaqCategoryEntityResponse>;
  Content?: Maybe<Scalars['String']['output']>;
  Title?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  locale?: Maybe<Scalars['String']['output']>;
  localizations?: Maybe<IgloohomeFaqRelationResponseCollection>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};


export type IgloohomeFaqLocalizationsArgs = {
  filters?: InputMaybe<IgloohomeFaqFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type IgloohomeFaqCategory = {
  __typename?: 'IgloohomeFaqCategory';
  Description?: Maybe<Scalars['String']['output']>;
  Name?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  locale?: Maybe<Scalars['String']['output']>;
  localizations?: Maybe<IgloohomeFaqCategoryRelationResponseCollection>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};


export type IgloohomeFaqCategoryLocalizationsArgs = {
  filters?: InputMaybe<IgloohomeFaqCategoryFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type IgloohomeFaqCategoryEntity = {
  __typename?: 'IgloohomeFaqCategoryEntity';
  attributes?: Maybe<IgloohomeFaqCategory>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type IgloohomeFaqCategoryEntityResponse = {
  __typename?: 'IgloohomeFaqCategoryEntityResponse';
  data?: Maybe<IgloohomeFaqCategoryEntity>;
};

export type IgloohomeFaqCategoryEntityResponseCollection = {
  __typename?: 'IgloohomeFaqCategoryEntityResponseCollection';
  data: Array<IgloohomeFaqCategoryEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type IgloohomeFaqCategoryRelationResponseCollection = {
  __typename?: 'IgloohomeFaqCategoryRelationResponseCollection';
  data: Array<IgloohomeFaqCategoryEntity>;
};

export type IgloohomeFaqEntity = {
  __typename?: 'IgloohomeFaqEntity';
  attributes?: Maybe<IgloohomeFaq>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type IgloohomeFaqEntityResponse = {
  __typename?: 'IgloohomeFaqEntityResponse';
  data?: Maybe<IgloohomeFaqEntity>;
};

export type IgloohomeFaqEntityResponseCollection = {
  __typename?: 'IgloohomeFaqEntityResponseCollection';
  data: Array<IgloohomeFaqEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type IgloohomeFaqRelationResponseCollection = {
  __typename?: 'IgloohomeFaqRelationResponseCollection';
  data: Array<IgloohomeFaqEntity>;
};
// ...
export type Query = {
  __typename?: 'Query';
  i18NLocale?: Maybe<I18NLocaleEntityResponse>;
  i18NLocales?: Maybe<I18NLocaleEntityResponseCollection>;
  igloohomeChangelog?: Maybe<IgloohomeChangelogEntityResponse>;
  igloohomeChangelogs?: Maybe<IgloohomeChangelogEntityResponseCollection>;
  igloohomeFaq?: Maybe<IgloohomeFaqEntityResponse>;
  igloohomeFaqCategories?: Maybe<IgloohomeFaqCategoryEntityResponseCollection>;
  igloohomeFaqCategory?: Maybe<IgloohomeFaqCategoryEntityResponse>;
  igloohomeFaqs?: Maybe<IgloohomeFaqEntityResponseCollection>;
  me?: Maybe<UsersPermissionsMe>;
  search?: Maybe<SearchResponse>;
  uploadFile?: Maybe<UploadFileEntityResponse>;
  uploadFiles?: Maybe<UploadFileEntityResponseCollection>;
  uploadFolder?: Maybe<UploadFolderEntityResponse>;
  uploadFolders?: Maybe<UploadFolderEntityResponseCollection>;
  usersPermissionsRole?: Maybe<UsersPermissionsRoleEntityResponse>;
  usersPermissionsRoles?: Maybe<UsersPermissionsRoleEntityResponseCollection>;
  usersPermissionsUser?: Maybe<UsersPermissionsUserEntityResponse>;
  usersPermissionsUsers?: Maybe<UsersPermissionsUserEntityResponseCollection>;
};
// ...
export type ResponseCollectionMeta = {
  __typename?: 'ResponseCollectionMeta';
  pagination: Pagination;
};

export type SearchResponse = {
  __typename?: 'SearchResponse';
  igloohomeFaqs?: Maybe<IgloohomeFaqEntityResponseCollection>;
};
// ...
export type UploadFileEntityResponse = {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
};

export type UploadFileEntityResponseCollection = {
  __typename?: 'UploadFileEntityResponseCollection';
  data: Array<UploadFileEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type UploadFolderEntityResponse = {
  __typename?: 'UploadFolderEntityResponse';
  data?: Maybe<UploadFolderEntity>;
};

export type UploadFolderEntityResponseCollection = {
  __typename?: 'UploadFolderEntityResponseCollection';
  data: Array<UploadFolderEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type UsersPermissionsMe = {
  __typename?: 'UsersPermissionsMe';
  blocked?: Maybe<Scalars['Boolean']['output']>;
  confirmed?: Maybe<Scalars['Boolean']['output']>;
  email?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  role?: Maybe<UsersPermissionsMeRole>;
  username: Scalars['String']['output'];
};
// ...
export type UsersPermissionsRoleEntityResponse = {
  __typename?: 'UsersPermissionsRoleEntityResponse';
  data?: Maybe<UsersPermissionsRoleEntity>;
};

export type UsersPermissionsRoleEntityResponseCollection = {
  __typename?: 'UsersPermissionsRoleEntityResponseCollection';
  data: Array<UsersPermissionsRoleEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type UsersPermissionsUserEntityResponse = {
  __typename?: 'UsersPermissionsUserEntityResponse';
  data?: Maybe<UsersPermissionsUserEntity>;
};

export type UsersPermissionsUserEntityResponseCollection = {
  __typename?: 'UsersPermissionsUserEntityResponseCollection';
  data: Array<UsersPermissionsUserEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type CategoriesQuery = { __typename?: 'Query', igloohomeFaqCategories?: { __typename?: 'IgloohomeFaqCategoryEntityResponseCollection', data: Array<{ __typename?: 'IgloohomeFaqCategoryEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaqCategory', Name?: string | null } | null }> } | null };
// ...
export type ChangelogsQuery = { __typename?: 'Query', igloohomeChangelogs?: { __typename?: 'IgloohomeChangelogEntityResponseCollection', data: Array<{ __typename?: 'IgloohomeChangelogEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeChangelog', AppVersion?: string | null, ReleaseDate?: any | null, Changes?: string | null } | null }> } | null };
// ...
export type FaqsQuery = { __typename?: 'Query', igloohomeFaqs?: { __typename?: 'IgloohomeFaqEntityResponseCollection', data: Array<{ __typename?: 'IgloohomeFaqEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaq', Title?: string | null, Content?: string | null, Category?: { __typename?: 'IgloohomeFaqCategoryEntityResponse', data?: { __typename?: 'IgloohomeFaqCategoryEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaqCategory', Name?: string | null } | null } | null } | null } | null }> } | null };